move server to API
